requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.1.0
python-docx==1.1.0
pdfminer.six==20221105
pytesseract==0.3.10
reportlab==4.2.2
markdown==3.5.1
requests==2.32.3